% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_post.R
\name{create_post}
\alias{create_post}
\title{Create a post from the template}
\usage{
create_post(title = "A study of Zika and Cholera", slug = NULL,
  category = c("tutorials", "practicals", "case studies"),
  author = c("Celina Turchi", "John Snow"), date = Sys.Date(),
  licence_name = "CC-BY",
  licence_url = "https://creativecommons.org/licenses/by/3.0/")
}
\arguments{
\item{title}{Title for the post}

\item{slug}{Short informative slug with words separated by "-"}

\item{category}{Either "tutorials", "practicals" or "case studies"}

\item{author}{Author as a vector}

\item{date}{Date for the post}

\item{licence_name}{Name of the license}

\item{licence_url}{URL to the license}
}
\value{
Used only for side-effects.
}
\description{
Create a post from the template
}
\details{
create a R Markdown file and open it.
}
