% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_axis_line.R
\name{draw_axis_line}
\alias{draw_axis_line}
\title{draw_axis_line}
\usage{
draw_axis_line(length_x, length_y, tick_step = NULL, lab_step = NULL)
}
\arguments{
\item{length_x}{the length of x axis, a numeric}

\item{length_y}{the length of y axis, a numeric}

\item{tick_step}{the length of tick's step, a numeric}

\item{lab_step}{the length of label's step, a numeric}
}
\value{
an object of ggplot that we can add ggplot layer by "+"
}
\description{
The source code from the url:
\url{ https://stackoverflow.com/questions/17753101/center-x-and-y-axis-with-ggplot2}
}
\examples{
p <- draw_axis_line(20, 4)
library(ggplot2)
p + geom_point(aes(x =1, y =1))
}
