% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flat_violin.R
\name{geom_flat_violin}
\alias{geom_flat_violin}
\title{geom_flat_violin}
\usage{
geom_flat_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "dodge",
  trim = TRUE,
  scale = "area",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{trim}{whether trim the tails of the violins to the range of the data.}

\item{scale}{if "area" (default), all violins have the same area (before trimming the tails). If "count", areas are scaled proportionally to the number of observations. If "width", all violins have the same maximum width.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{whether inherit aesthetic mapping from `ggplot`}

\item{...}{additional parameters}
}
\value{
One side flat of geom_violin in ggplot2
}
\description{
The source code from the solution to:

\url{https://twitter.com/EamonCaddigan/status/646759751242620928}

The url of the source code is:

\url{https://gist.githubusercontent.com/benmarwick/2a1bb0133ff568cbe28d/raw/fb53bd97121f7f9ce947837ef1a4c65a73bffb3f/geom_flat_violin.R}

This function based mostly on copy/pasting from ggplot2 geom_violin source:

\url{https://github.com/hadley/ggplot2/blob/master/R/geom-violin.r}
}
