% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_and_update_dependencies.R
\name{get_and_update_dependencies}
\alias{get_and_update_dependencies}
\title{Determine all dependencies in blog posts and install them}
\usage{
get_and_update_dependencies(dir = "content/post/")
}
\arguments{
\item{dir}{Directory to search for packages in. Defaults to the blog post location of Hugo.}
}
\value{
Used for pure side effect
}
\description{
Using the automagic package, detect requirements in Rmd files
and install them all. Will help you if packages appear to be on GitHub.
}
