% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_new_rmds_to_md.R
\name{render_new_rmds_to_md}
\alias{render_new_rmds_to_md}
\title{Prep Hugo pages by converting Rmds to Md's}
\usage{
render_new_rmds_to_md(dir = "content/post", build = "old and new")
}
\arguments{
\item{dir}{Directory to search for new Rmds in. Defaults to the blog post location of Hugo.}

\item{build}{Selection criteria for which Rmds to convert to md, options are c("all", "old and new", "old", "new")}
}
\value{
Used for pure side effect
}
\description{
Assumes anything without a corresponding markdown file
needs to be renderd to md. Hugo (the static site generator)
then converts these md files to html on the site. Use the
preview option in Rstudio to quickly see how your md has worked.
}
