% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_use_widget.R
\name{save_and_use_widget}
\alias{save_and_use_widget}
\title{Store a widget and generate iframe code}
\usage{
save_and_use_widget(x, filename)
}
\arguments{
\item{x}{htmlwidget to be shown}

\item{filename}{Unique filename for the widget to be saved under}
}
\value{
HTML code for iframe of htmlwidget
}
\description{
Use this to embed a htmlwidget in your Rmd's for when they
get rendered to markdown. You also need to add 
`always_allow_html: yes` to the Rmd for the html to remain
incorporated. Widget will be saved in the widget directory.
}
