% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_level.R
\name{sign_level}
\alias{sign_level}
\title{sign_level}
\usage{
sign_level(
  data,
  sig.level = c(0.001, 0.01, 0.05, 0.1),
  mark = c("***", "**", "*", ".")
)
}
\arguments{
\item{data}{matrix}

\item{mark}{}
}
\value{
return statistical significance mark
}
\description{
we want to determine the significance levels of the p-value by
this function
}
