% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_chull.R
\docType{data}
\name{StatChull}
\alias{StatChull}
\title{stat_chull}
\usage{
StatChull
}
\arguments{
\item{mapping}{Set of aesthetic mappings.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer.}

\item{inherit.aes}{whether inherit aesthetic mapping from `ggplot`}

\item{...}{additional parameters}
}
\value{
give the convex hull of a set of points in ggplot2.
}
\description{
The source code from the documents of

official extension mechanism provided in ggplot2

\url{https://cran.r-project.org/web/packages/ggplot2/vignettes/extending-ggplot2.html}
}
\keyword{datasets}
