% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmc.R
\name{wmc}
\alias{wmc}
\title{wmc}
\usage{
wmc(formula, data, exact = FALSE, sort = TRUE, method = "holm")
}
\value{
Nonparametric pairwise multiple comparisons using the Wilcoxon Signed Rank Test,
and probability values are adjusted using the p.adjust function.
}
\description{
Nonparametric pairwise multiple comparisons using the Wilcoxon Signed Rank Test

Probability values are adjusted using the p.adjust function

The url of the source code is:

\url{https://www.statmethods.net/RiA/wmc.txt}
}
